requirements.txt
annotated-types==0.6.0
anyio==4.3.0
asgiref==3.7.2
attrs==23.1.0
beautifulsoup4==4.12.2
browser-cookie3==0.19.1
cachetools==5.3.3
certifi==2023.7.22
cffi==1.16.0
charset-normalizer==3.3.1
colorama==0.4.6
comtypes==1.2.1
configparser==6.0.0
crayons==0.4.0
cryptography==41.0.7
Django==5.0.1
django-autoslug==1.9.9
django-tinymce==3.7.1
google-ai-generativelanguage==0.6.2
google-api-core==2.18.0
google-api-python-client==2.127.0
google-auth==2.29.0
google-auth-httplib2==0.2.0
google-cloud-vision==3.7.2
google-generativeai==0.5.2
googleapis-common-protos==1.63.0
grpcio==1.62.2
grpcio-status==1.62.2
h11==0.14.0
h2==4.1.0
hpack==4.0.0
httpcore==1.0.5
httplib2==0.22.0
httpx==0.27.0
hyperframe==6.0.1
idna==3.4
install==1.3.5
loguru==0.7.2
lz4==4.3.3
MouseInfo==0.1.3
numpy==1.26.2
opencv-python==4.8.1.78
outcome==1.3.0
packaging==23.2
Pillow==10.1.0
proto-plus==1.23.0
protobuf==4.25.3
pyasn1==0.6.0
pyasn1_modules==0.4.0
PyAudio==0.2.14
PyAutoGUI==0.9.54
pycparser==2.21
pycryptodomex==3.20.0
pydantic==2.7.1
pydantic_core==2.18.2
PyGetWindow==0.0.9
PyMsgBox==1.0.9
pyOpenSSL==23.3.0
pyparsing==3.1.2
pyperclip==1.8.2
pypiwin32==223
PyQt5==5.15.10
PyQt5-Qt5==5.15.2
PyQt5-sip==12.13.0
PyRect==0.2.0
PyScreeze==0.1.30
PySocks==1.7.1
python-dotenv==1.0.0
python-gemini-api==2.4.2
pyttsx3==2.90
pytweening==1.0.7
pywin32==306
requests==2.31.0
rsa==4.9
selenium==4.6.0
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.5
SpeechRecognition==3.10.1
sqlparse==0.4.4
tqdm==4.66.2
trio==0.22.2
trio-websocket==0.11.1
typing_extensions==4.9.0
tzdata==2023.4
uritemplate==4.1.1
urllib3==1.26.18
urllib3-secure-extra==0.1.0
webdriver-manager==4.0.1
wikipedia==1.4.0
win32-setctime==1.1.0
wsproto==1.2.0
